iommu: msm: Only register int. for non-secure CB
If a page fault occurs the IOMMU interrupt handler calls the IOMMU page
fault handler for each context to check which context caused the error.
However, only the secure environment has access to the registers for
secure context and thus the IOMMU kernel driver should not add handler
for this context to avoid illegal access.

CRs-fixed: 444537
Change-Id: I4f7cbe3959282f29d24e123b3d092929a917285a
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>