media: video: Verify vsync sequence error is real
Due to a hw bug it is possible for a vsync sequence error to be
falsely triggered. A sw workaround that waits to see what the
next error interrupt is after a vsync seq err will help
indentify if the issue is a false positive.

CRs-Fixed: 429178
Signed-off-by: Terence Hampson <thampson@codeaurora.org>
(cherry picked from commit 7188b7fdfcc9027836ae75ed8628f08cb1edc25d)

Change-Id: Ie3319bf7d7028579c0cf16f118ded199e23aac7f
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>