msm: clock-8610: Preserve behavior of the MDP core clock
The MDP core clock used on previous targets has been
dropped; the MDP AXI clock also serves as the core clock.
In order to preserve previous behavior, add a voter clock
that will allow the MDP driver to set a rate on the core_clk
as has been done for previous targets.

Change-Id: Id475fff5b912d88f16cd22676cee66f355b90468
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>