CHROMIUM: power: Allow a power supply to be located by device node
Power supply naming depends on the attached bus and device. For example
an SBS device might have a name like sbs-004-1b, where the characters after
"sbs" come from the i2c bus to which the battery is attached.

This is not really suitable for device tree operation. We want to use a
phandle to point to the power supply, rather than hard-coded knowledge of
what linux might decide to name it.

Add a function to permit a power supply to be located by device node, which
is easy to discover from a phandle.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
See that charger-manager can locate the sbs battery correctly now when
using the device tree.

Change-Id: I0c906469e84406879285669c44f00224eee30136
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43014
Reviewed-by: Benson Leung <bleung@chromium.org>