msm_serial_hs : Add support to configure GPIOs
This change adds support to configure/unconfigure GPIOs connected
to HSUART peripheral if GPIOs related information is provided with
enumerated HSUART port. GPIOs would be configured with active
configuration when HSUART port is opened and would be configured
with suspend configuration when HSUART port is closed by application
provided GPIOMUX settings are available for those GPIOs with platform
specific configuration.

Change-Id: I9f7f0ae11535d9c2eded69729bdb2942fc907190
Signed-off-by: Saket Saurabh <ssaurabh@codeaurora.org>