power: qpnp-bms: fix current and cc raw conversion
Currently, the BMS is wrongly using nanovolts as the conversion
unit for the ideal IADC gain. Change this to raw.

Also, the BMS is using the gain read from the IADC driver to convert
the current and coulomb counter when it should be using an offset
gain. Fix this by subtracting the offset from gain in the BMS driver
before doing the conversion.

Change-Id: Ib47d70fee798ab50a557cae1556afc8f662fa331
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>