Revert "nohz: Fix idle ticks in cpu summary line of /proc/stat"
This reverts commit 7386cdbf2f57ea8cff3c9fde93f206e58b9fe13f.

Commit 21111be8 (ARM: Fix negative idle stats for offline cpu) changed
offline cpus to be not seen in nohz mode (their ts->idle_active is 0).
With that change, we don't need the check for cpu_online() introduced in
commit 7386cdbf.  Offline cpu's idle time as last recorded in its
ts->idle_sleeptime will be reported (thus excluding its offline time as
part of idle time statistics). Having fewer code sites worry about online
status of cpus is a good thing!

Note: This revert is currently applicable only for ARM architecture. This
revert will cause cpu idle statistics to be "more" broken on other
architectures, unless they are modified to have a change similar to
21111be8.

Change-Id: I03bb599e9a8abf4a56802288c31ba5a2a93a8647
Signed-off-by: Srivatsa Vaddagiri <vatsa@codeaurora.org>