CHROMIUM: dark resume: Make rtc_cmos part of kernel
This change makes the rtc_cmos driver part of the kernel instead of a module.
This allows us to modify sysfs files for it (such as dark_resume_source) as part
of an upstart script. We need to change the owner of the dark_resume_source file
to power so powerd can enable the rtc cmos as a wake source.

BUG=chrome-os-partner:9812
TEST=check that rtc cmos is probed before user space starts.

Change-Id: Ie562fe78ade4f6b46ee1de970401435f52200fa4
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42926
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>