CHROMIUM: rtc: Add a method to obtain the wake interrupt number
When an rtc wakes the system we need to find this out so that we can decide
whether to immediately suspend again, or to fully wake to user space. This
is required for drivers that want to monitor devices preiodically in suspend,
such as charger-manager.

Add a new rtc method which returns the IRQ used by an RTC. This IRQ number
can then be checked in the suspend_again handler, to see if it caused the
wake.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
Tested with charger-manager with other changes. See that it can remain
suspended (i.e. returns to suspend automatically after each wake).

Change-Id: I885aac2dfe15a3fb2c420857382224b98d475416
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43019
Reviewed-by: Benson Leung <bleung@chromium.org>