hwmon: qpnp-adc: Add synchronous vadc and iadc api
Support simultaneous voltage and current ADC request
to read vbatt and battery current. The IADC peripheral
issues the simultaneous VADC/IADC sync request. The
VADC peripheral is first prepared to issue an voltage
ADC request but does not request the conversion. The
IADC peripheral sets its conversion request and issues
an ADC request for simultaneous read. The api is
intended to be used by the BMS.

Change-Id: I978db99662313d7344bd2e4c8534f23cab398011
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>