drm/exynos: Untangle dpms a little
This patch strives to simplify the various dpms paths. The overarching
goal is to not get into a situation where connector->dpms is out of
touch with the rest of the driver since that effectively disables
userspace control of the power. The patch also eliminates the duplicate
paths through crtc and encoder.

BUG=chromium-os:37797
TEST=Tested hotplug; idle suspend/resume; idle off/on; lid
     suspend/resume; powerd_suspend/resume. No regressions noted

Change-Id: I79376e07500633505eb0159edac9a4ed9054b0e4
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43196