ASoC: Avoid putting stream state to STOP when stream paused
When multiple Front-end are using the same Back-end, putting
the stream state for a Front-end to STOP will make the the Backend
stream to be released, when the front-end is paused. This will avoid
the Backend to be released (when another active front-end stream is
present).
Instead, put the stream state to PAUSED, which will keep the Back-end
ON.

Change-Id: Ic31bf90a09b605b91f6b9fe3c09361b1145581fb
CRs-fixed: 450610
Signed-off-by: Banajit Goswami <bgoswa@codeaurora.org>
(cherry picked from commit 9d2004e0448cbe5f6199510fa9999699b4a0a574)

Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>