drm/exynos: Don't reset mixer on win_disable
This patch removes code which resets the mixer on win_disable of the
default overlay. The effect of this code is that it disables interrupts,
and resets all mixer registers to power-on-reset values, then
re-enables interrupts. The code was introduced in the S2R patch, it
probably hacked around a bug, but it's not needed any longer.

BUG=None
TEST=on snow; powerd_suspend/idle off/idle suspend/hotplug/multiple boots

Change-Id: Id65650c7ab7a4cbc508a2087d4354b780acdce1a
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43517
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>