CHROMIUM: drm/exynos: dp: configure GPIO for eDP hotplug detection
the HPD line from the Parade eDP/LVDS bridge is used to trigger the
display controller initialization. Let's configure it as a proper
interrupt line to allow the Exynos display port controller driver to use
it.
The current Exynos needs special hackery to plug an interrupt on a
random GPIO.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:17228
TEST=on Spring, boot and see the counter of the HPD interrupt from
the eDP bridge incrementing.

Change-Id: I15c612fe57c0dab98e18c58e1ef092fd8e103a92
Reviewed-on: https://gerrit.chromium.org/gerrit/43535
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>