drm/exynos: Remove crtc->disable
Remove the crtc->disable callback since it's not needed and most likely
incorrect anyways. DRM will fall back to dpms if disable is not defined,
and that will simplify the various paths we have. Furthermore, all
disable did was disable the default overlay on the crtc, which is not
the intended result.

BUG=None
TEST=Tested on snow with boot/hotplug/suspend/idle

Change-Id: I2b134662bfa56d35a0de9fbf761d7a8a3708009e
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43725