CHROMIUM: TEMPORARY: instrumentation for crosbug.com/p/17609
We have a hard time reproducing crosbug.com/p/17609.  It seems
to be correlated with a reset of the device.  We don't understand
why the reset happens and we're hoping that a stack trace will
give us some information.

From the logs we've seen, this doesn't happen very often, so
chances of spewing are low.  We get Link feedback reports with
this error almost daily.

BUG=chrome-os-partner:17609
TEST=not tested, or, rather, YOU test it
BRANCH=none

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: If24b4f2490919905d79cdf4e5b8b159cbe8aee6c
Reviewed-on: https://gerrit.chromium.org/gerrit/43832
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>