CHROMIUM: cpufreq: interactive: allow arbitrary speed / delay mappings
Accept a string of delays and speeds at which to apply the delay before
raising each step above hispeed. For example, "80000 1300000:200000
1500000:40000" means that the delay at or above 1GHz, until 1.3GHz is 80 msecs,
the delay until 1.5GHz is 200 msecs and the delay at or above 1.5GHz is 40
msecs when hispeed_freq is 1GHz.

[toddpoynor@google.com: add documentation]
Signed-off-by: Minsung Kim <ms925.kim@samsung.com>
(cherry picked from android-3.4 commit
	e43dde3b6cec1e1d36a477713bcfb1dda070bef2)

BUG=chrome-os-partner:20830
TEST=Boot Pit; able to set hispeed delay values.  For example:
$ echo "20000 700000:0 1600000:20000" > above_hispeed_delay
sets the delay to 20ms when going from 650Mhz to 700Mhz, 0ms
when scaling between 700Mhz to 1.6Ghz, and 20ms when scaling
above 1.6Ghz.

Change-Id: If9f519250d1edc697b0c2f8497551219139a40a6
[abrestic: fixed conflicts due to not having target_loads]
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63415
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66245