msm: ipc_logging: add compile time check on ipc_log_string()
ipc_log_string accepts printf style formatted strings and arguments.  Add
the gcc __printf() attribute to ipc_log_string() for compile time checking
of the currect number and types of the arguments based on the provided
format string to prevent stack corruption or other errors due to an
incorrect combination of format string and arguments provided to
ipc_log_string().

Change-Id: I20c290df0f4083d4342f6b48c3008119eafcd0d8
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>