msm: clock-8974: Remove get_parent_byte() and get_parent_pixel()
Neither set_rate_byte() nor set_rate_pixel() are written to handle
clk->parent being NULL, and they will crash if this happens. Since
these clocks will only ever be used with one parent
(dsipll0_pixel_clk_src), just hard-code the parent pointer to that.

The source selection for this parent is written into registers as
part of set_rate_hid() when it is called from set_rate_byte() and
set_rate_pixel(), so this assumed parent is guaranteed to be
correct by the time the clock is enabled.

Change-Id: Iafb02020234e3a9e1273425224f2b6a703fe38c8
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>