CHROMIUM: Input: atkbd - skip re-probing on non-KBD i8042 ports
The serio0 or "i8042 KBD port" (defined in i8042.c) is used for a system to
connect AT keyboard device in general. When a system resumes from suspend,
kernel rescans all serio ports again for matching devices with serio
driver one by one. Currently, the retry mechanism in atkbd_probe() may take
few milliseconds for each serial port. There will be four auxiliary ports if
multiplexing is enabled on the second serial port, and this will add the
latency of system resume or boot significantly. The patch tries to skip
reprobe mechanism for non-KBD ports in order to reduce the unresponsiveness
for a system with ps/2-type of touchpad resuming from suspend. By skipping
reprobes on non-KBD i8042 ports it could save 1+ seconds of latency.

BUG=chrome-os-partner:15015
TEST=on device; use powerd_suspend and then hit a key to wake it up
   see if the trackpoint/trackpoint could be active in 2+ seconds
   after resume instead of 3~4 seconds long

Change-Id: I0e62c351225b06af44b55e8600f6e37b05ae5a06
Signed-off-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44518
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>