CHROMIUM: mfd: init/link earlier
Move multi function devices earlier in the link sequence, in parity with
regulators.
The eDP bridge loaded by the DRM driver needs to be able to control the
screen regulators.

The regulators linking was moved earlier to have them available for
other subsystems :
https://lkml.org/lkml/2008/12/2/10

Now a lot of regulators are instantiated from MFD, so we need to move
the MFD to keep this benefit.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:18122
TEST=boot on Spring and see the traces from the virtual TPS65090
regulators earlier than the display port initialization.

Change-Id: I62223ea9fa79651a787754d45e05dd26edb7a029
Reviewed-on: https://gerrit.chromium.org/gerrit/44598
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>