power: pm8921-bms: restrict soc decrease to one percent
The calculated soc could decrease up to 3% within one
calculation period. This will happen when UUC and CC increase
by 1% percent while the adjusting algorithm reduces the OCV
to cause RC to decrease by 1%.

Update the reporting side of the code to restrict the decrease
of soc by 1%. There will be ample time to catch up to calculated
soc.

Importantly, when the device is resumed after a long time, we don't
want to restrict the soc decrease. So when the device goes to suspend
forget the last soc reported so the next soc report is unrestricted.

Change-Id: I100516237ee9ab4176ca29c067f094db200652f8
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>