CHROMIUM: rtc: fix bug with setting control register
A bit in the RTC_CONTROL register was accidentally lost in cmos_caused_wake.
It seems that this effectively moved us from 24 hour format to 12 hour format in
practice. This can cause the time from the RTC to be misinterpreted, so the
system will read the wrong time from the RTC. It might then go back to an
old value for the system time (and also store that in the RTC).

BUG=chromium:216936
TEST=suspend resume system (with dark resume happening)

Change-Id: I1425b857caf49ce3e538dfc05b981153729565f9
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Previous-Reviewed-on: https://gerrit.chromium.org/gerrit/45823
(cherry picked from commit ff64c149db61f2b35c98678ec2b0192a3207302d)
Reviewed-on: https://gerrit.chromium.org/gerrit/46024
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>