gpu: ion: Don't have a mapping for secure buffers
Secure buffers can't have a cpu mapping. The CPU may
speculate in to any mapping in the page tables. This
may result in security violations from speculative accesses.
Adjust the allocation to support this. Additionally,
ensure that any function which might try to use a cpu
mapping are restricted.

Change-Id: I85c7fd77c62519d6afa423d2fc125c03a5e2ca78
CRs-Fixed: 465059
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>