ehci-msm2: configure MPM interrupt to wakeup target on usb remote wakeup
During XO shutdown, USB can detect the remote wakeup signal and generate
ASYNC interrupt to QGIC, which in turn propogates this interrupt to MPM.
But currently MPM is not configured to wakeup the KRAIT on receiving this
interrupt. Configure this MPM to KRAIT interrupt, so MPM can wakeup
the target.

CRs-Fixed: 481792
Change-Id: I47b9fcf2ecf3d07d9180be6a2c58aaac4e13336a
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>