msm: Kconfig: remove dependencies and incorrect defaults for SENSORS_ADSP
The config SENSORS_ADSP can be selected by future targets and has
no particular dependency on 8974 or 8226. The dependency list will
keep increasing as more targets use this config. Its more suitable
to have individual targets select it in their defconfigs instead.
This config does not need to be enabled by default either. This was
introduced in ab00c019bfb7b0af19a4341759b3ac381a168edf as an unwanted
side effect since the actual intention of the change was to remove it
from the Kconfig.

Change-Id: Id613156b61038dd58bbe5336c17d8ab5c76ad987
Signed-off-by: Syed Rameez Mustafa <rameezmustafa@codeaurora.org>