tools/power: fspin -- a utility for power measurements
fspin is a simple utility for generating a controlled workload
for use when testing various Linux power management features.
In particular, for testing the various flavors of power capping.

When measuring the benefit of these features,
it is important to simultaneously measure not only
the power benefit, but the performance impact.

See the man page for details.

Signed-off-by: Len Brown <len.brown@intel.com>