msm: pil-q6v5-lpass: Turn on clocks when using secure ops
Certain clocks required for LPASS operation are turned on
when the adsp is brought out of reset and turned off after
the adsp is shutdown in the non-secure ops. The secure world
replicates this behavior.

Unfortunately, some of these clocks may be used by other
clients in Linux. Therefore, turn on/off these clocks in
the secure ops as well.

Change-Id: Ifd7da7e0c9ad8958cf001db86a389d7c3ece722c
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>