msm: kgsl: better handling of virtual address fragmentation
When KGSL_MEMFLAGS_USE_CPU_MAP is enabled, the mmap address
must try to match the GPU alignment requirements of the buffer,
as well as include space in the mapping for the guard page.
This can cause -ENOMEM to be returned from get_unmapped_area()
when there are a large number of mappings. When this happens,
fall back to page alignment and retry to avoid failure.

Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>
(cherry picked from commit 85593353ab237454c5bff82746be8fcdae148c6d)

Change-Id: I3122df38c960ae4be59d3745d5a2b51414dae363
Signed-off-by: Shashank Babu Chinta Venkata <sbchin@codeaurora.org>