BACKPORT: drm/i915: make intel_ddi_get_cdclk_freq return values in KHz
With this, that 338 can finally become the correct 337500.

Due to the change we need to adjust the intel_dp_aux_ch function to
set the correct value, so adjust the division and also use
DIV_ROUND_CLOSEST instead of the old "round down" behavior because the
spec says the value "should be programmed to get as close as possible
to the ideal rate of 2MHz".

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
(cherry picked from commit 71067e8fcf4d769c0714b75e4c1b4220ec467638)

BUG=chrome-os-partner:19682
TEST=With entire patch series applied, observe PC7 residency via powertop in
     crosh.
Change-Id: If861fda4f07b15ca23ff0a9590e6e449b4145eea
Reviewed-on: https://gerrit.chromium.org/gerrit/56482
Tested-by: Josh Triplett <josh@joshtriplett.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>