netconsole: s/syslogd/cancd/ in documentation
Using syslogd to capture netconsole is known to be broken, see for
example https://bugzilla.redhat.com/show_bug.cgi?id=432160 or any of the
many other bug reports.  We should not advertise it, much less as a
first choice.  The fact that syslogd tends to initially work makes it
worse, as that creates false hope.

Cancd is a syslog-for-netconsole of sorts and in my experience works
better than any alternative for non-trivial setups, i.e. more than a
single machine sending netconsole traffic.

Since my hacked-up version of cancd is no longer compatible with
Oracle's original, I linked to both.

Signed-off-by: Joern Engel <joern@logfs.org>