msm: vidc: ensure max capabilities for vp8 on 8x10
For vp8 decode on 8x10, the max height and width
is less than the default values. The input buffer
allocation is done based on the max capabilities
values sent by firmware, but the buffer plane
sizes are set as per the default values. This
results in qbuf failures. This change ensures
that max capabilities are not overwritten for 8x10.

Change-Id: I0018e06e7d7c0dc4de821665ed3ca6145b600f56
Signed-off-by: Arun Menon <avmenon@codeaurora.org>