CHERRY-PICK: UPSTREAM: mmc: block: Enable runtime pm for mmc blkdevice
Once the mmc blkdevice is being probed, runtime pm will be enabled.
By using runtime autosuspend, the power save operations can be done
when request inactivity occurs for a certain time. Right now the
selected timeout value is set to 3 s. Obviously this value will likely
need to be configurable somehow since it needs to be trimmed depending
on the power save algorithm.

For SD-combo cards, we are still leaving the enablement of runtime PM
to the SDIO init sequence since it depends on the capabilities of the
SDIO func driver.

Moreover, when the blk device is being suspended, we make sure the device
will be runtime resumed. The reason for doing this is that we want the
host suspend sequence to be unaware of any runtime power save operations
done for the card in this phase. Thus it can just handle the suspend as
the card is fully powered from a runtime perspective.

Finally, this patch prepares to make it possible to move BKOPS handling
into the runtime callbacks for the mmc bus_ops. Thus IDLE BKOPS can be
accomplished.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Chris Ball <cjb@laptop.org>
(cherry picked from commit e94cfef698aae6b209d8918dd319312e4b02118d)

BUG=chrome-os-partner:29136
TEST=Continuously cold reboot Blaze; no eMMC corruption observed.

Original Change-Id: I319d531201aa79ac8a616d9a9f92790f661e5826
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202368
Reviewed-by: Dylan Reid <dgreid@chromium.org>

Change-Id: I9f54f34d56e971f7c6100f0686e56ab7d21b0e99
Reviewed-on: https://chromium-review.googlesource.com/210801
Commit-Queue: Star Lin <star_lin@compal.com>
Tested-by: Star Lin <star_lin@compal.com>
Reviewed-by: Mao Huang <littlecvr@chromium.org>