CHROMIUM: gpio: tegra: moving suspend function to syscore
When supporting deep sleep mode on Tegra which means except always-on
power partitions and SDRAM in self-refresh mode, other blocks and main
power will be shut off. The GPIO settings can't be retained across power
cycle. It will cause the resume flow be broken when some of the early
resume functions or IRQ accessing the device that depends on GPIO to work
normally. Moving the suspend/resume function from system to syscore to
resolve this.

BUG=none
TEST=test with LP0/LP1 suspend-resume

Change-Id: I18b17d3da276565638f7793883000658b7a5e6cc
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/176993
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>