CHROMIUM: gpio: tegra: moving suspend function to syscore
When supporting deep sleep mode on Tegra which means except always-on
power partitions and SDRAM in self-refresh mode, other blocks and main
power will be shut off. The GPIO settings can't be retained across power
cycle. It will cause the resume flow be broken when some of the early
resume functions or IRQ accessing the device that depends on GPIO to work
normally. Moving the suspend/resume function from system to syscore to
resolve this.

BUG=chrome-os-partner:40741
TEST=LP0 suspend/resume test

Change-Id: Ieff2e3dc1cd4d6d334a4df3ef00e0c968ebec559
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/276943
Reviewed-by: Benson Leung <bleung@chromium.org>