qup_i2c: Add vote for clock path
Currently the i2c driver votes for its core clock. However, it is
possible that the core-clock's source will be off. When the source
is off, the i2c core is not clocked. This patch adds voting for
the entire clock path, thus guarantee clocking to the i2c core.
Voting for the path cost a few msecs. A client may choose when
to pay this overhead. By default the votes takes place in the
runtime-pm callbacks which tie to transaction requests. A client
may choose to invoke these callbacks itself whenever it likes
by directly calling the driver's runtime-pm. A third option is
to set the active_only flag in platform data which results in the
callbacks being called when the application processor power state
changes.

CRs-Fixed: 483873
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>
(cherry picked from commit 0906336a050fa47118be4fddf7a7d2dcf21a3986)

Change-Id: I74cca9054fddf817e9d1d98fdb1c2a9790624b28
Signed-off-by: Sridhar Nelloru <nellos@codeaurora.org>