media: video: Allow for vcap vc to be paused and resumed
There are some instances where the client of vcap does not want
to call stream off and set up everything again for vcap to stream
again. This patch allows for vcap vc to be paused. While vcap vc
is paused it is possible for client to give new timing values,
as long as there is no change to the size of buffers needed.

Change-Id: I628d29cdc87ac41f4f7cbed76511cfc7f8475fac
Signed-off-by: Terence Hampson <thampson@codeaurora.org>