CHROMIUM: i2c-designware-pci: Add Haswell ULT device IDs
Add the necessary PCI Device IDs to use the Haswell ULT
I2C controller in PCI mode.

Ideally we would use the controller in ACPI mode, but that
is a messy backport for 3.8.

I am not 100% positive on the configuration for haswell,
but based on how it was done for the ACPI mode driver these
are same right settings.  I set the bus numbers to -1 so
it would use dynamic assignment rather than hardcoded.

BUG=chrome-os-partner:19495
TEST=manual: successfully load i2c-designware-pci driver

Change-Id: Ia81c0329ac946af7d78018e67acb18a9e5ef8697
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/51375
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>