CHERRY-PICK: CHROMIUM: Disable runtime PM suspend for quirky xhci controllers.
If a USB controller with XHCI_RESET_ON_RESUME goes to runtime suspend,
a reset will be performed upon runtime resume. Any previously suspended
devices attached to the controller will be re-enumerated at this time.
This will cause problems, for example, if an open system call on the
device triggered the resume (the open call will fail).

Note that this change is only relevant when persist_enabled is not set
for USB devices.

BUG=chromium:239214
TEST=Manual. Verify on XHCI_RESET_ON_RESUME platform (Stout) that USB
controller doesn't go to runtime suspend and USB webcam continues to
function in Chrome.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>

Change-Id: I0cc8d0cbbcef27c3f8a708086af1b7c5cd897c7c
Original-Change-Id: I89206cc1427d9a94d8e7ac912126d45c803dde3b
Reviewed-on: https://gerrit.chromium.org/gerrit/51378
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56606