UPSTREAM: ARM: mm: correct pte_same behaviour for LPAE.
For 3 levels of paging the PTE_EXT_NG bit will be set for user
address ptes that are written to a page table but not for ptes
created with mk_pte.

This can cause some comparison tests made by pte_same to fail
spuriously and lead to other problems.

To correct this behaviour, we mask off PTE_EXT_NG for any pte that
is present before running the comparison.

Signed-off-by: Steve Capper <steve.capper@linaro.org>
Reviewed-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit dde1b65110353517816bcbc58539463396202244)

BUG=chrome-os-partner:25909
TEST=4GB Norrin boots with DVFS and display disabled.

Change-Id: I485bf8ba96054db03440bdadfc578d2b0636d064
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187610