arm/dt: msm8610: Add active-set only and sleep-set only regulators for L10
LDO 10 is used by the acpuclock on msm8610. However it is currently
configured as active+sleep. Since the sleep request isn't different,
rpm-regulator-smd does not send it.

Fix this by adding active-set only and sleep-set only regulator nodes
for LDO 10. Configure the sleep-set only node so that it is disabled
by default. Change the consumer mapping for acpuclock so that the new
active-set only node is used instead of the old active+sleep node.

Change-Id: Id0b66ddf1623364e20b37090f53bb01c98dfa7c7
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>