msm: clock-local2: Check if clock is enabled before setting rate
Add extra checking for cases where a clock's prepare count are zero but
the clock is already enabled; i.e. from boot rom.  If this condition is
true, enable the source clock to ensure we don't switch to one that is
disabled.  Doing so will fail to update the rcg clock and can lead to
bus lockups during early boot.

Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>