CHROMIUM: workaround for Infineon TPM broken defensive timeout
The SLB9836 TPM by Infineon goes into a self-defense mode when power
is cut in the middle of a command too many times, and a bug prevents
it from coming out of that mode in a reasonable time frame.  This
patch adds a shutdown-time callback that grabs the mutex wrapping a
request-response transaction, so it's less likely that we'll lose
power in the middle of a command.

Note that the TPM may lose power also across suspend/resume (depending
on hardware).  However, the suspend callback sends a SaveState command,
which can only be sent when other running commands have completed.
I am not sure if a command can start after the suspend callback
has completed (I doubt it) but even if it does, it will be
a sufficiently rare event, and the self-defense can tolerate that.

BUG=chromium:234796
TEST=extensively tested with the librarian simulator
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: If1424f931f3da2c7d259c7d56050ab68f94f1e5a
Reviewed-on: https://gerrit.chromium.org/gerrit/57988
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58406