CHERRY-PICK: HACK: Always print the wake source on resume
This exposes the print_active_wakeup_sources function from wakeup.c and
calls it fom the wakeup path in suspend.c for debugging purposes, this
allows one to see what most likely woke the system.

This is a cherry pick from https://gerrit.chromium.org/gerrit/#/c/58003/
which is unlikely to land in the kernel, but will be useful for debugging
spurious wakes on the factory line.

BUG=chrome-os-partner:19958
TEST=Manual, run on DUT, powerd_suspend, check log for message.

Change-Id: I330ac81963e0d8968ed08990b36fc2d7342ca3ab
Original-Change-Id: If707807068387df2f88d946e31873a23853a341f
Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58696
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Queue: Jon Salz <jsalz@chromium.org>
Tested-by: Jon Salz <jsalz@chromium.org>