CHROMIUM: tg3: Wait for boot code to finish after power on
Some systems that don't need wake-on-lan may choose to power down the
chip on system standby. Upon resume, the power on causes the boot code
to startup and initialize the hardware. On one new platform, this is
causing the device to go into a bad state due to a race between the
driver and boot code, once every several hundred resumes. The same race
exists on open since we come up from a power on.

This patch adds a wait for boot code signature at the beginning of
tg3_init_hw() which is common to both cases. If there has not been a
power-off or the boot code has already completed, the signature will be
present and poll_fw() returns immediately. Also return immediately if
the device does not have firmware.

BUG=chrome-os-partner:18649
TEST=suspend/resume testing

Change-Id: Ia1413299695fd8eb7f3400d20a9b38c290f22a66
Original-Change-Id: I761cff4cdfb524ea8fdbb2bad6880c7d98f78b85
Signed-off-by: Ashwin Thiagarajan <ashwint@broadcom.com>
Signed-off-by: Mohammed Habibulla <moch@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58367
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58560