cfg80211: Add "interface unavailable" event
NL80211_CMD_INTERFACE_UNAVAILABLE command is used to indicate that
the interface is not available to userspace. The driver may have
encountered errors that has forced the driver to report this event.

When the driver is in P2P-Go mode, and when it goes thru automatic
recovery from firmware crashes, it uses this interface to notify the
userspace that the group has been deleted. The userspace should retry
once this event is received.

Change-Id: I9e5f40e8382617f3aeeaf0a3d76e004c869fbfad
Signed-off-by: Sameer Thalappil <sameert@codeaurora.org>