exynos: Slow down firing of the watchdog on boot
There is already code to make the watchdog wait for quite a while. With
the reset value for WTCON on Pit we get about a second between setting
up the watchdog and needing to reset it.

With earlyprintk this is not always long enough. So change the divisor
from 16 to 128. This gives us about 8 seconds which should be plenty.

Also move the enable code to after this initial setup, since it seems
that there there is a small chance that we might reboot if the watchdog
happens to count to 0 just as we enable reset. The time window would be
1/32768 seconds or about 30us.

BUG=chrome-os-partner:20181
TEST=manual
Boot kernel with these bootargs:

setenv bootargs "cros_legacy console=ttySAC3,115200 debug slub_debug=FZPUA root=/dev/mmcblk0p3 rootwait ro earlyprintk"

See that it no longer resets in the middle of the GPIO printouts.

Change-Id: I2096ab95d1bdcf4dff77e37fa6eb06524718ecb6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58697
Reviewed-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 3f8fd28c409b41d6bc2b4c069ee80f4e549bd6d3)
Reviewed-on: https://gerrit.chromium.org/gerrit/59506
Commit-Queue: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Katie Roberts-Hoffman <katierh@chromium.org>